requirements.txt
deep_translator==1.11.4
discord.py==2.4.0
python-dotenv==1.0.1
python_Levenshtein==0.26.1
python_Levenshtein==0.26.1
python_magic==0.4.27
youtube_search_python==1.6.6
yt_dlp==2025.2.19
